pyproject.toml
[project]
name = "ai-blog-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cohere>=5.16.1",
    "dnspython>=2.7.0",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "pymongo>=4.13.2",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
